requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
h11==0.14.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
